requirements.txt
# Artisans Hub - Full-Stack Web Application
# Core Flask Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Pillow==10.0.0
opencv-python-headless==4.8.1.78
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.19

# Image Processing & AI (simplified for production)
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76

# File Upload & Storage
# secure-filename is part of werkzeug

# AI Mock/Simulation (for demo purposes)
# random, json, datetime are built-in modules

# API utilities
requests==2.31.0
python-dotenv==1.0.0

# Development tools
pytest==7.4.2